tweak propelauth frontend and move on to backend
