+++ Pipfile
@@ -13,10 +13,10 @@ flask-sqlalchemy = "3.0.3"
 flask-migrate = "*"
 sqlalchemy-serializer = "*"
 Werkzeug = "2.2.2"
-
 marshmallow = ">3.0.0, <4.0.0"
 flask-marshmallow = "*"
 marshmallow-sqlalchemy = "*"
+propelauth-flask = "*"
 
 [requires]
 python_full_version = "3.8.13"
