+++ client/src/auth/LogOut.js
@@ -1,0 +1,17 @@
+import React from 'react';
+
+
+// Local imports
+import Header from '../layout/Header';
+
+function LogOutPage() {
+
+  return (
+    <div>
+      <Header />
+      You are now logged out.
+    </div>
+  );
+}
+
+export default LogOutPage;
