+++ client/src/pages/NotFoundPage.js
@@ -1,7 +1,10 @@
 import React from "react";
 import { Link, useLocation } from "react-router-dom";
+
+// Local imports
 import Search from "../components/Search";
 import Footer from "../layout/Footer";
+import NotFoundImage from "../common/404.jpg"
 
 function NotFoundPage() {
     const location = useLocation();
@@ -11,11 +14,12 @@ function NotFoundPage() {
         <div>
             <h1>Oops! You tried to visit "{location.pathname}".</h1>
             <p>But it might have been moved, deleted, or perhaps never existed.</p>
-            <img src="404_image.jpg" alt="Page not found" />
             <div>
                 <Link to="/">Return to Home</Link>
             </div>
             <p>Or use this search bar:</p><Search />
+            <img src={NotFoundImage} alt="Page not found" />
+
             <Footer />
         </div>
     );
